+++ common/src/main/java/kg/delletenebre/yamus/api/response/Playlist.kt
@@ -28,7 +28,7 @@ data class Playlist(
     val textColor: String = "",
     val title: String = "",
     val trackCount: Int = 0,
-    val uid: Int = 0,
+    val uid: Long = 0,
     val visibility: String = ""
 ) {
     @Serializable
@@ -46,7 +46,7 @@ data class Playlist(
         val login: String = "",
         val name: String = "",
         val sex: String = "",
-        val uid: Int = 0,
+        val uid: Long = 0,
         val verified: Boolean = false
     )
