+++ common/src/main/java/kg/delletenebre/yamus/api/response/PlaylistIds.kt
@@ -3,4 +3,4 @@ package kg.delletenebre.yamus.api.response
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class PlaylistIds(val uid: Int, val kind: Int)
+data class PlaylistIds(val uid: Long, val kind: Int)
